requirements.txt
asgiref==3.4.1
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
cycler==0.11.0
distro==1.6.0
Django==4.0.1
django-cors-headers==3.10.1
djangorestframework==3.13.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
fonttools==4.28.5
greenlet==1.1.2
idna==3.3
jsonschema==3.2.0
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.0
packaging==21.3
pandas==1.3.5
paramiko==2.9.2
Pillow==9.0.0
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.29
sqlparse==0.4.2
texttable==1.6.4
urllib3==1.26.8
websocket-client==0.59.0